+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ### Added
 - Add documentation in the style of [Read the Docs](https://readthedocs.org/).
 
+### Changed
+- `dp compile`:
+  - `--env` option has a default value: `local`,
+  - `--datahub` is changed to `--datahub-gms-uri`, `--repository` is changed to `--docker-repository-uri`.
+- `dp deploy`'s `--docker-push` is not a flag anymore and requires a Docker repository URI parameter; `--repository` got removed then.
+
 ## [0.6.0] - 2021-12-16
 
 ### Modified